requirements.txt
uiautomation
Pillow
pywin32
psutil
pyperclip
win11toast
pystray